fix: add correct work under tests

+++ src/routes/profiles/index.ts
@@ -2,7 +2,17 @@ import { FastifyPluginAsyncJsonSchemaToTs } from '@fastify/type-provider-json-sc
 import { idParamSchema } from '../../utils/reusedSchemas';
 import { createProfileBodySchema, changeProfileBodySchema } from './schema';
 import type { ProfileEntity } from '../../utils/DB/entities/DBProfiles';
-import ErrorMsg from '../types';
+
+enum ErrorMsg {
+  profile_not_found = 'Profile not found',
+  profile_already_exist = 'Profile already exist',
+  member_type_not_found = 'Member type not found'
+};
+
+enum ErrorCode {
+  code_400 = 400,
+  code_404 = 404
+};
 
 const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
   fastify
@@ -23,7 +33,7 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       if (profile) {
         return profile;
       }
-      throw new Error(ErrorMsg.profile_not_found);
+      throw fastify.httpErrors.createError(ErrorCode.code_404, ErrorMsg.profile_not_found);
     }
   );
 
@@ -35,7 +45,15 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      return await fastify.db.profiles.create(request.body);
+      const memberType = await fastify.db.memberTypes.findOne({key: 'id', equals: request.body.memberTypeId});
+      if (memberType) {
+        const profile = await fastify.db.profiles.findOne({key: 'userId', equals: request.body.userId});
+        if (profile) {
+          throw fastify.httpErrors.createError(ErrorCode.code_400, ErrorMsg.profile_already_exist);
+        }
+        return await fastify.db.profiles.create(request.body);
+      }
+      throw fastify.httpErrors.createError(ErrorCode.code_400, ErrorMsg.member_type_not_found);
     }
   );
 
@@ -47,7 +65,11 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      return await fastify.db.profiles.delete(request.params.id);
+      const profile = await fastify.db.profiles.findOne({key: 'id', equals: request.params.id});
+      if (profile) {
+        return await fastify.db.profiles.delete(request.params.id);
+      }
+      throw fastify.httpErrors.createError(ErrorCode.code_400, ErrorMsg.profile_not_found);
     }
   );
 
@@ -60,7 +82,11 @@ const plugin: FastifyPluginAsyncJsonSchemaToTs = async (
       },
     },
     async function (request, reply): Promise<ProfileEntity> {
-      return await fastify.db.profiles.change(request.params.id, request.body);
+      const profile = await fastify.db.profiles.findOne({key: 'id', equals: request.params.id});
+      if (profile) {
+        return await fastify.db.profiles.change(request.params.id, request.body);
+      }
+      throw fastify.httpErrors.createError(ErrorCode.code_400, ErrorMsg.profile_not_found);
     }
   );
 };
